% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.collapse_polygon}
\alias{.collapse_polygon}
\title{Collapses the original polygon by linking each segment to the last vertex
of the previous segment, basically connecting all sides to the base.}
\usage{
.collapse_polygon(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The coordinates of the collapsed polygon as a data frame.
}
\description{
Collapses the original polygon by linking each segment to the last vertex
of the previous segment, basically connecting all sides to the base.
}
\examples{
.collapse_polygon(data.frame(x=c(8, 8, 22, 24), y=c(12, 15, 8, 8)))

}
\seealso{
\code{\link[=.get_area]{.get_area()}}
}
