% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-lines.R
\name{.draw_mean_lines}
\alias{.draw_mean_lines}
\title{Plots the mean lines}
\usage{
.draw_mean_lines(df, mean_line_color)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}

\item{mean_line_color}{color of the mean line inside the polygon}
}
\value{
The mean lines as a \code{ggplot2} object.
}
\description{
Plots the mean lines
}
\seealso{
\code{\link[=.get_perimeter]{.get_perimeter()}}
}
