% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_lengths}
\alias{.get_lengths}
\title{Calculates the length of each side of the polygon}
\usage{
.get_lengths(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The list of lengths as a numeric vector.
}
\description{
Calculates the length of each side of the polygon
}
\examples{
.get_lengths(data.frame(x=c(8, 8, 22, 24), y=c(12, 15, 8, 8)))

}
\seealso{
\code{\link[=.get_perimeter]{.get_perimeter()}}
}
