% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-polygon.R
\name{.get_mean_coords}
\alias{.get_mean_coords}
\title{Calculates the}
\usage{
.get_mean_coords(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The list of lengths as a numeric vector.
}
\description{
Calculates the
}
\examples{
.get_lengths(data.frame(x=c(8, 8, 22, 24), y=c(12, 15, 8, 8)))

}
\seealso{
\code{\link[=.get_perimeter]{.get_perimeter()}}
}
