% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{prepare_dataframe}
\alias{prepare_dataframe}
\title{Prepare dataframe for polygon plot}
\usage{
prepare_dataframe(dataframe)
}
\arguments{
\item{dataframe}{classic data.frame with numeric values}
}
\value{
\code{data.frame} object ready for the polygon plot function
}
\description{
Prepare dataframe for polygon plot
}
\examples{
classic_df <- data.frame(
  data1=c(-300, 500, 300, 350),
  data2=c(4, 14, -2, 12),
  data3=c(0, 1000, 10, 50),
  data4=c(0, 50, 20, 40)
)

df <- prepare_dataframe(classic_df)
df
}
