% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.get_area}
\alias{.get_area}
\title{Calculates the area of the polygon.}
\usage{
.get_area(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The area as numeric.
}
\description{
Calculates the area of the polygon.
}
\seealso{
\code{\link[=.collapse_polygon]{.collapse_polygon()}}
}
