% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_area}
\alias{.get_area}
\title{Calculates the area of the polygon. It serves as a wrapper for the different
functions for each polygon type.}
\usage{
.get_area(df, type = "square")
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The area as float.
}
\description{
Calculates the area of the polygon. It serves as a wrapper for the different
functions for each polygon type.
}
\examples{
.get_area(data.frame(x=c(8, 8, 22, 24), y=c(12, 15, 8, 8)), type="square")

}
\seealso{
\code{\link[=.get_area_square]{.get_area_square()}}

\code{\link[=.get_perimeter]{.get_perimeter()}}
}
