% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_area_square}
\alias{.get_area_square}
\title{Calculates the area of the square.}
\usage{
.get_area_square(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The area as float.
}
\description{
Calculates the area of the square.
}
\examples{
.get_area(x=x, y=y)

}
\seealso{
\code{\link[=.get_area]{.get_area()}}
}
