% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-lines.R
\name{.get_mean_coords}
\alias{.get_mean_coords}
\title{Calculates the mean coordinates for further plotting the mean polygon}
\usage{
.get_mean_coords(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
A data.frame with the mean coordinates.
}
\description{
Calculates the mean coordinates for further plotting the mean polygon
}
\seealso{
\code{\link[=.get_perimeter]{.get_perimeter()}}
}
