% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_perimeter}
\alias{.get_perimeter}
\title{Calculates the perimeter of the polygon}
\usage{
.get_perimeter(df)
}
\arguments{
\item{df}{Data frame containing the coordinates in two columns: x and y}
}
\value{
The perimeter as float.
}
\description{
Calculates the perimeter of the polygon
}
\examples{
.get_perimeter(x=x, y=y)

}
\seealso{
\code{\link[=.get_area]{.get_area()}}
}
