% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_polygons.R
\name{overlay_polygons}
\alias{overlay_polygons}
\title{Overlay different polygon plots on the same axis}
\usage{
overlay_polygons(
  plot_list,
  label_list,
  legend_title = "Overlay Polygons",
  include_mean_lines = FALSE
)
}
\arguments{
\item{plot_list}{is a list object containing a series of polygon plots
(ggplot2 object). All the polygon plot should be drawn on the same axis range}

\item{label_list}{is a vector of strings. Must have the same length of
\code{plot_list} object}

\item{legend_title}{title of the legend}

\item{include_mean_lines}{logical flag indicating whether to include mean
lines in the overlaid polygons}
}
\value{
a ggplot2 object
}
\description{
Overlay different polygon plots on the same axis
}
