% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonplot.R
\name{polygonPlot}
\alias{polygonPlot}
\title{Draw polygon plot}
\usage{
polygonPlot(
  df,
  shape,
  extra = 0.5,
  fillcolor = "black",
  alpha = 0.5,
  linecolor = "black",
  linetype = "solid",
  lwd = 0.8,
  labels_axis = NULL,
  mean_line_color = "darkgrey",
  annotation_tick_size = 3.5,
  annotation_label_size = 5,
  title = NULL,
  fix_aspect_ratio = TRUE
)
}
\arguments{
\item{df}{data.frame containing the data. The first column of the dataframe
must contain the strings \code{axis_min}, \code{axis_max} and \code{data} in that order.
\code{axis_min} and \code{axis_max} are for the axis ranges. If left empty, the
axis_min/max will be automatically calculated based on data given and
extension will be decided based on the Axes range extension
(\code{extra} parameter).
From the second column to the last, they need to contains the value
of \code{axis_min}, \code{axis_max}, and \code{data} for all the variables of interest.
From the third row, each cell can contain data values for the corresponding
column variable. Column names will be used as labels of the axis
(starting from the second one). Change them in order to change the labels,
but don't leave them empty.}

\item{shape}{integer value to specify the shape of the polygon (3=Triangle,
4=Square, 5=Pentagon, 6=Hexagon)}

\item{extra}{axis range extension}

\item{fillcolor}{fill color of the polygon}

\item{alpha}{alpha value of the fill color}

\item{linecolor}{line color of the polygon border}

\item{linetype}{line type of the polygon border}

\item{lwd}{line width of the polygon border}

\item{labels_axis}{vector with the desired labels of the axis}

\item{mean_line_color}{color of the mean line inside the polygon}

\item{annotation_tick_size}{size of tick marks along the polygon axes}

\item{annotation_label_size}{size of the labels displayed along the polygon axes}

\item{title}{title of the plot}

\item{fix_aspect_ratio}{Boolean flag to fix the aspect ratio of the plot as
\code{1}. It is strongly recommended to leave it as default value \code{TRUE}.
NOTE: If you are going to change the theme of the returned ggplot object,
remember to put in the \code{theme} function the following code
\code{aspect.ratio = 1} in order to keep the text and the relative
ticks aligned on the axis.}
}
\value{
A \code{ggplot2} object.
}
\description{
Draw polygon plot
}
\examples{
df <- data.frame(
  info=c("axis_min", "axis_max", "data", ""),
  data1=c(-300, 500, 300, 350),
  data2=c(4, 14, -2, 12),
  data3=c(0, 1000, 10, 50),
  data4=c(0, 50, 20, 40)
)

p <- polygonPlot(df, shape=4, fillcolor = "dodgerblue", linecolor = "blue")
p 

}
